Organized some code, fixed issues and generilizations of the task system, and began work on clients.

+++ server/src/locations.rs
@@ -2,8 +2,8 @@ pub const CONFIG_PATH: &str = "/etc/regis/regis.conf";
 pub const TOTAL_DIR: &str = "/etc/regis";
 
 /// Represents the default console port used by regis.
-pub const CONSOLE_PORT: u32 = 1025;
+pub const CONSOLE_PORT: u16 = 1025;
 /// Represents the default hosts port used by regis.
-pub const HOSTS_PORT: u32 = 1026;
+pub const HOSTS_PORT: u16 = 1026;
 /// Represents the default broadcast port used by regis.
-pub const BROADCAST_PORT: u32 = 1027;
+pub const BROADCAST_PORT: u16 = 1027;
